I am now finished with the assigment's coding part. This is not exactly my desired outcome, but I understand why the program behaves the way it does. I could have made the AI's in pairs so that they always play against the same opponent. That would maybe give better results. Or maybe not. Crossover would still make the result a bit funky.
The last option that may improve the AI would be to have it find the optimal square each time a move were made, with the current state of the game in mind. This way I would be able to play against it as well.

Another issue I see with this optimization algorithm is that is does not care if it is the minimizing or maximizing player, like the minimax algorithm does.